+++ client/client.ts
@@ -96,6 +96,7 @@ export class Client {
   // CodeMirror editor
   editorView!: EditorView;
   commandKeyHandlerCompartment?: Compartment;
+  vimCompartment?: Compartment;
   indentUnitCompartment?: Compartment;
   undoHistoryCompartment?: Compartment;
 
@@ -822,8 +823,12 @@ export class Client {
         ) {
           this.versionMismatchNotified = true;
           this.ui.flashNotification(
-            "A new version of SilverBullet is available. Reload to update.",
+            "A new version of SilverBullet client is available. A reload or two is required to update.",
             "warning",
+            {
+              timeout: 0,
+              actions: [{ name: "Reload", run: () => location.reload() }],
+            },
           );
         }
         break;